requirements.txt
requests
polyline
opencv-python
matplotlib
numpy
torch
transformers
pillow
flask
